♻️ [#110] Refactor: useEscToClose 훅 분리
